Split out util.RapidWright into StandaloneEntrypoint
